treat rename events like deletes to avoid F_GETPATH
FreeBSD doesn't support F_GETPATH, so the existing
handler for rename events does not work on FreeBSD (the fcntl()
call fails with EINVAL); in fact, FreeBSD does not appear to
provide any way to determine the path associated with an open
file descriptor. Since renames will always cause a modification
(and thus trigger a re-scan) of the new parent directory, we can
simply treat them as deletes; with that, a rename or move will
generate a pair of delete and add events rather than a pair of
rename and add events as the current implementation does.